move some operations into their own submodule
