+++ crates/noirc_evaluator/src/ssa/acir_gen/operations.rs
@@ -1,0 +1,5 @@
+pub mod binary;
+mod bitwise;
+mod cmp;
+
+pub mod constrain;
